Develop

See merge request s05-webmobile1-sub3/S05P13B308!162

+++ backend/src/main/java/com/ssafy/square4us/api/mvc/service/MemberServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ssafy.square4us.api.mvc.service;
 
+import com.ssafy.square4us.api.mvc.model.dto.FileDTO;
 import com.ssafy.square4us.api.mvc.model.dto.MemberDTO;
 import com.ssafy.square4us.api.mvc.model.entity.FileEntity;
 import com.ssafy.square4us.api.mvc.model.entity.Member;
@@ -7,16 +8,17 @@ import com.ssafy.square4us.api.mvc.model.entity.MemberRole;
 import com.ssafy.square4us.api.mvc.model.repository.FileRepository;
 import com.ssafy.square4us.api.mvc.model.repository.MemberRepository;
 import com.ssafy.square4us.api.mvc.model.repository.MemberRepositorySupport;
+import com.ssafy.square4us.api.mvc.model.repository.StudyMemberRepository;
+import com.ssafy.square4us.common.util.S3Util;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
 /**
  * ReadOnly 달고 Create, Update, Delete는 따로 표기
@@ -28,13 +30,17 @@ public class MemberServiceImpl implements MemberService {
     private final String PROFILE_PATH = System.getProperty("user.dir") + "\\profile";
 
     private final MemberRepository memberRepository;
+    private final StudyMemberRepository studyMemberRepo;
     private final MemberRepositorySupport memberRepositorySupport;
     private final FileRepository fileRepository;
+    private final S3Util s3Util;
 
-    public MemberServiceImpl(MemberRepository memberRepository, MemberRepositorySupport memberRepositorySupport, FileRepository fileRepository) {
+    public MemberServiceImpl(MemberRepository memberRepository, StudyMemberRepository studyMemberRepo, MemberRepositorySupport memberRepositorySupport, FileRepository fileRepository, S3Util s3Util) {
         this.memberRepository = memberRepository;
+        this.studyMemberRepo = studyMemberRepo;
         this.memberRepositorySupport = memberRepositorySupport;
         this.fileRepository = fileRepository;
+        this.s3Util = s3Util;
     }
 
     @Override
@@ -82,16 +88,13 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     @Transactional
-    public Long updateMemberByEmail(String email, MemberDTO.UpdatePatchReq updateInfo) {
-        Member member = Member.builder().email(email)
-                .nickname(updateInfo.getNickname())
-                .build();
-        return memberRepositorySupport.updateByMemberEmail(member);
+    public Long updateMemberByEmail(Long memberId, MemberDTO.UpdatePatchReq updateInfo) {
+        return memberRepositorySupport.updateByMemberEmail(memberId, updateInfo);
     }
 
     @Override
-    @Transactional
-    public MemberDTO updateProfileByEmail(String email, MultipartFile profile) {
+    @Transactional(rollbackFor = IOException.class)
+    public MemberDTO updateProfileByEmail(String email, MultipartFile profile) throws IOException {
         Optional<Member> find = memberRepository.findByEmail(email);
         if(!find.isPresent()) {
             return null;
@@ -99,10 +102,7 @@ public class MemberServiceImpl implements MemberService {
         Member member = find.get();
         deletePrevProfile(member);
 
-        int code = uploadProfile(member, profile);
-        if(code == 1) {
-            return null;
-        }
+        uploadProfile(member, profile);
 
         return new MemberDTO(member);
     }
@@ -126,19 +126,26 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     @Transactional
-    public void deleteMemberByEmail(String email) {
+    public Boolean deleteMemberByEmail(String email) {
+        Optional<Member> member = memberRepository.findByEmail(email);
+        if(!member.isPresent()){
+            return false;
+        }
+
+        if(studyMemberRepo.existsLeaderJoinStudy(member.get().getId())!=0){
+            return false;
+        }
+
         deleteProfileByEmail(email);
         memberRepository.deleteByEmail(email);
+        return true;
     }
 
     @Transactional
     public MemberDTO deletePrevProfile(Member member) {
         FileEntity prevProfile = member.getProfile();
         if(prevProfile != null) {
-            File del = new File(prevProfile.getFilePath(), prevProfile.getFileName());
-            if(del.exists()) {
-                del.delete();
-            }
+            s3Util.delete(prevProfile);
             fileRepository.deleteById(prevProfile.getId());
         }
         member.setProfile(null);
@@ -146,36 +153,24 @@ public class MemberServiceImpl implements MemberService {
     }
 
     @Transactional
-    public int uploadProfile(Member member, MultipartFile profile) {
-        File path = new File(PROFILE_PATH);
-        if(!path.exists()) {
-            path.mkdir();
-        }
-        String originName = profile.getOriginalFilename();
-        String contentType = profile.getContentType();
-        String uuid = UUID.randomUUID().toString();
-        String saveFileName = uuid + originName.substring(originName.lastIndexOf('.'));
-
-        FileEntity file = FileEntity.builder()
-                .member(member)
-                .study(null)
-                .article(null)
-                .meeting(null)
-                .filePath(PROFILE_PATH)
-                .fileName(saveFileName)
-                .fileOriginName(originName)
-                .contentType(contentType)
-                .build();
-
-        file = fileRepository.save(file);
-
+    public void uploadProfile(Member member, MultipartFile profile) throws IOException {
+        String path = "profile";
         try {
-            profile.transferTo(new File(PROFILE_PATH, saveFileName));
-        } catch (IOException e) {
-            return 1;
-        }
-        member.setProfile(file);
+            FileDTO fd = s3Util.upload(profile, path);
+            FileEntity file = FileEntity.builder()
+                    .member(member)
+                    .article(null)
+                    .meeting(null)
+                    .filePath(fd.getFilePath())
+                    .fileName(fd.getFileName())
+                    .fileOriginName(fd.getFileOriginName())
+                    .contentType(fd.getContentType())
+                    .build();
 
-        return 2;
+            file = fileRepository.save(file);
+            member.setProfile(file);
+        } catch (IOException e) {
+            throw new IOException("파일 저장 실패!");
+        }
     }
 }
